moved RPCClient into client file
